Move R2R settings to r2r module
