+++ src/r2r.rs
@@ -1,6 +1,16 @@
-use super::R2R_KIND;
+use super::{FLAG, R2R_KIND};
+use super::aligned_vec::*;
+use super::error::*;
+use super::pair::{Pair, ToPair};
+use super::plan::R2R;
 
-pub fn forward(kind: R2R_KIND) -> R2R_KIND {
+use ffi;
+
+use ndarray::*;
+use num_traits::Zero;
+use std::marker::PhantomData;
+
+fn forward(kind: R2R_KIND) -> R2R_KIND {
     match kind {
         R2R_KIND::FFTW_R2HC => R2R_KIND::FFTW_R2HC,
         R2R_KIND::FFTW_HC2R => R2R_KIND::FFTW_R2HC,
@@ -16,7 +26,7 @@ pub fn forward(kind: R2R_KIND) -> R2R_KIND {
     }
 }
 
-pub fn backward(kind: R2R_KIND) -> R2R_KIND {
+fn backward(kind: R2R_KIND) -> R2R_KIND {
     match kind {
         R2R_KIND::FFTW_R2HC => R2R_KIND::FFTW_HC2R,
         R2R_KIND::FFTW_HC2R => R2R_KIND::FFTW_HC2R,
@@ -33,7 +43,7 @@ pub fn backward(kind: R2R_KIND) -> R2R_KIND {
 }
 
 /// see http://www.fftw.org/fftw3_doc/Real_002dto_002dReal-Transform-Kinds.html
-pub fn logical_size(n: usize, kind: R2R_KIND) -> usize {
+fn logical_size(n: usize, kind: R2R_KIND) -> usize {
     match kind {
         R2R_KIND::FFTW_R2HC => n,
         R2R_KIND::FFTW_HC2R => n,
@@ -46,5 +56,46 @@ pub fn logical_size(n: usize, kind: R2R_KIND) -> usize {
         R2R_KIND::FFTW_RODFT01 => 2 * n,
         R2R_KIND::FFTW_RODFT10 => 2 * n,
         R2R_KIND::FFTW_RODFT11 => 2 * n,
+    }
+}
+
+#[derive(Debug, Clone, Copy, new)]
+pub struct R2R1D {
+    n: usize,
+    kind: R2R_KIND,
+    flag: FLAG,
+}
+
+pub fn r2hc_1d(n: usize) -> R2R1D {
+    R2R1D {
+        n: n,
+        kind: R2R_KIND::FFTW_R2HC,
+        flag: ffi::FFTW_MEASURE,
+    }
+}
+
+impl<T: R2R + AlignedAllocable + Zero> ToPair<T, T> for R2R1D {
+    type Dim = Ix1;
+    fn to_pair(&self) -> Result<Pair<T, T, Ix1>> {
+        let mut field = AlignedVec::new(self.n);
+        let mut coef = AlignedVec::new(self.n);
+        let forward = unsafe { T::r2r_1d(self.n, &mut field, &mut coef, forward(self.kind), self.flag) };
+        let backward = unsafe {
+            T::r2r_1d(
+                self.n,
+                &mut coef,
+                &mut field,
+                backward(self.kind),
+                self.flag,
+            )
+        };
+        Pair {
+            field: field,
+            coef: coef,
+            logical_size: logical_size(self.n, self.kind),
+            forward: forward,
+            backward: backward,
+            phantom: PhantomData,
+        }.null_checked()
     }
 }
